.gitignore, Terrain Tools and trees removal.
Added to .gitignore, added Terrain Tools and removed distant trees because of weird LOD effect.
Also edited the water color to be darker, in order for it to be visible from a distance through the fog, and adjusted the colour of the wood materials a little to look darker and more realistic.