Chopping Wood Progress
Probuilder has been used to split imported wood model into pieces, which are then used to make prefabs in which the orginal model is broken into pieces.  A healthsystem script has been written, with the intend of using it for a wood chopping script. The wood chopping script is in progress