Changed the Shader for many materials
Changed the Shader for materials on the buildings, the vegetation and props in the scene.
The 'Complex Lit' shader that was enabled before gave everything (especially the wood buildings), a too bright and glossy look which was very prevalent in scene 3 because it is a very dark scene.

Also changed the SMOOTHNESS of many of the materials in the scene for buildings and vegetation to make it look more life-like.