NPC talk focus switching
Implemented the feature of being able to switch which NPC you want to talk to, based on Gaze.
If you look at another NPC for 3 seconds, they will be the currently selected NPC to talk to and the previous NPC will have your chat log stored on them.

Still need to implement a switching of the audio source and change the voice based on currently selected NPC.