Gaze interaction button 'CharacterTesting' scene
Think I fixed so you can press a button by gazing at it first and then pressing the primary button on the left controller.

Just added an 'XR Controller' script component on the 'Gazeinteractor' gameobject, and a input binding for the UI press action.
Together with the XR Ray interactor, the XR controller should enable the input actions to register.