NPC focus audio switching
Added a way for switching NPC character focus when gazing that ALSO changes the audio source now and the neural TTS voice.

Still needs tweaking, since the chatlog stored on the NPC which is being set to the ChatGPT prompt, overrides itself (I think) since the NPC is not able to finish their sentences.