Optimized ChatGPT executions and Game World Info
- The VR UI that showed the chatlog, together with all the code in ChatTest.cs that tinkered with this UI has now been deprecated. This should optimize performance and remove some warnings about the SendMessage method.
 Instead look in console to see chatlog now.

- Scene 3 is all set up for multiple NPC dialogue.

- Added some basic game world story and game world info in the inspector under the 'World Info' gameobject which the NPCs use for reference. This is the added text:

Game Story: "A funeral is taking place today for a man called Torsten. Everyone from the village are helping prepare for the funeral. The Adventurer is a person who is passing through the village but has decided to help the villagers prepare for the funeral in exchange for a bed to sleep in for the night."

Game World: "You are a viking villager from the 900th century living in a small village next to a burial hill named Lindholm Høje located in the northern part of Jylland, Denmark."