New code for item description by ChatGPT
Added new code in the ChatTest.cs and NPCInteractorScript.cs that works with the new ChatGPT prompting setup, which enables the user to pick up the personal belongings in Scene 2 and make ChatGPT generate a descriptive dialogue of the item.

Currently only works with the horn since this is the only item that has been written a item description prompt for.