Object reset if out of bounds
Added a script that can be attached to gameobjects which will reset the object's position to the transform defined in a transform field for the script in the inspector, WHEN THE PLAYER IS NOT LOOKING AT THE OBJECT.
In coherence with this, I also added an 'Outside Zone Trigger' and 'ResetPosition' gameobject as children under the 'Moveable Area' gameobject in the relevant scenes.
Outside Zone Trigger object is for detecting when the gameobject goes out of bounds of the Moveable Area (or player area so to speak), and the ResetPosition is just the gameobject transform used for setting where the object should be repositioned to if out of bounds.

!!THIS STILL NEEDS TO BE SET UP FOR ALL RELEVANT OBJECTS IN THE CORRESPONDING SCENES!! but the script is there ready to use.