OpenAI update fix (system_fingerprint)
Fixed OpenAI API throwing an error because of a new needed member type of "system_fingerprint" in the OpenAI package, related to the ChatGPT model we use.

The IDE says that the change I made will not be saved because it is within a package file, but try testing it for yourselves and we'll see. I tried closing and opening the project after making the change, and it still worked, so it must have saved something... lol