Able to set XR origin position in each scene
Deprecated the old method of setting XROrigin by the position and rotation of it in the last scene.
This would have worked GREAT if we were able to do wireless testing, but since the institutions encryted internet does not allow us to connect the VR headset, then we have to do this method instead.